WIP: move survey schedule to modal
